LRCI-3783 Move addDownstreamBuilds & removeDownstreamBuilds into ParentBuild
