+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/BaseBuild.java
@@ -52,71 +52,6 @@ import org.json.JSONObject;
  */
 public abstract class BaseBuild implements Build {
 
-	public void addDownstreamBuilds(Map<String, String> urlAxisNames) {
-		final Build thisBuild = this;
-
-		List<Callable<Build>> callables = new ArrayList<>(urlAxisNames.size());
-
-		for (Map.Entry<String, String> urlEntry : urlAxisNames.entrySet()) {
-			String url = urlEntry.getKey();
-
-			try {
-				url = JenkinsResultsParserUtil.getLocalURL(
-					JenkinsResultsParserUtil.decode(url));
-			}
-			catch (UnsupportedEncodingException unsupportedEncodingException) {
-				throw new IllegalArgumentException(
-					"Unable to decode " + url, unsupportedEncodingException);
-			}
-
-			if (!hasBuildURL(url)) {
-				final String axisName = urlEntry.getValue();
-				final String buildURL = url;
-
-				Callable<Build> callable = new Callable<Build>() {
-
-					@Override
-					public Build call() {
-						try {
-							return BuildFactory.newBuild(
-								buildURL, thisBuild, axisName);
-						}
-						catch (RuntimeException runtimeException) {
-							if (!isFromArchive()) {
-								NotificationUtil.sendSlackNotification(
-									runtimeException.getMessage() +
-										"\nBuild URL: " + buildURL,
-									"ci-notifications", "Build Object Failure");
-							}
-
-							return null;
-						}
-					}
-
-				};
-
-				callables.add(callable);
-			}
-		}
-
-		ParallelExecutor<Build> parallelExecutor = new ParallelExecutor<>(
-			callables, true, getExecutorService());
-
-		downstreamBuilds.addAll(
-			parallelExecutor.execute(1000L * 60L * 60L * 3L));
-	}
-
-	@Override
-	public void addDownstreamBuilds(String... urls) {
-		Map<String, String> urlAxisNames = new HashMap<>();
-
-		for (String url : urls) {
-			urlAxisNames.put(url, null);
-		}
-
-		addDownstreamBuilds(urlAxisNames);
-	}
-
 	@Override
 	public void archive() {
 		archive(getArchiveName());
@@ -1992,11 +1927,6 @@ public abstract class BaseBuild implements Build {
 		System.out.println(getReinvokedMessage());
 
 		reset();
-	}
-
-	@Override
-	public void removeDownstreamBuild(Build build) {
-		downstreamBuilds.remove(build);
 	}
 
 	@Override
