+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/BaseParentBuild.java
@@ -1,16 +1,92 @@
 /**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-FileCopyrightText: (c) 2023 Liferay, Inc. https://liferay.com
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
 package com.liferay.jenkins.results.parser;
 
-import java.io.File;
+import java.io.UnsupportedEncodingException;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.concurrent.Callable;
 
 /**
  * @author Michael Hashimoto
  */
 public abstract class BaseParentBuild extends BaseBuild implements ParentBuild {
+
+	public void addDownstreamBuilds(Map<String, String> urlAxisNames) {
+		final Build thisBuild = this;
+
+		List<Callable<Build>> callables = new ArrayList<>(urlAxisNames.size());
+
+		for (Map.Entry<String, String> urlEntry : urlAxisNames.entrySet()) {
+			String url = urlEntry.getKey();
+
+			try {
+				url = JenkinsResultsParserUtil.getLocalURL(
+					JenkinsResultsParserUtil.decode(url));
+			}
+			catch (UnsupportedEncodingException unsupportedEncodingException) {
+				throw new IllegalArgumentException(
+					"Unable to decode " + url, unsupportedEncodingException);
+			}
+
+			if (!hasBuildURL(url)) {
+				final String axisName = urlEntry.getValue();
+				final String buildURL = url;
+
+				Callable<Build> callable = new Callable<Build>() {
+
+					@Override
+					public Build call() {
+						try {
+							return BuildFactory.newBuild(
+								buildURL, thisBuild, axisName);
+						}
+						catch (RuntimeException runtimeException) {
+							if (!isFromArchive()) {
+								NotificationUtil.sendSlackNotification(
+									runtimeException.getMessage() +
+										"\nBuild URL: " + buildURL,
+									"ci-notifications", "Build Object Failure");
+							}
+
+							return null;
+						}
+					}
+
+				};
+
+				callables.add(callable);
+			}
+		}
+
+		ParallelExecutor<Build> parallelExecutor = new ParallelExecutor<>(
+			callables, true, getExecutorService());
+
+		downstreamBuilds.addAll(
+			parallelExecutor.execute(1000L * 60L * 60L * 3L));
+	}
+
+	@Override
+	public void addDownstreamBuilds(String... urls) {
+		Map<String, String> urlAxisNames = new HashMap<>();
+
+		for (String url : urls) {
+			urlAxisNames.put(url, null);
+		}
+
+		addDownstreamBuilds(urlAxisNames);
+	}
+
+	@Override
+	public void removeDownstreamBuild(Build build) {
+		downstreamBuilds.remove(build);
+	}
 
 	protected BaseParentBuild(String url) {
 		super(url);
