+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/ParentBuild.java
@@ -1,12 +1,21 @@
 /**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-FileCopyrightText: (c) 2023 Liferay, Inc. https://liferay.com
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
 package com.liferay.jenkins.results.parser;
 
+import java.util.Map;
+
 /**
  * @author Michael Hashimoto
  */
-public interface ParentBuild {
+public interface ParentBuild extends Build {
+
+	public void addDownstreamBuilds(Map<String, String> urlAxisNames);
+
+	public void addDownstreamBuilds(String... urls);
+
+	public void removeDownstreamBuild(Build build);
+
 }
