+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/TopLevelBuild.java
@@ -15,7 +15,7 @@ import org.dom4j.Element;
 /**
  * @author Michael Hashimoto
  */
-public interface TopLevelBuild extends Build {
+public interface TopLevelBuild extends ParentBuild {
 
 	public String getAcceptanceUpstreamJobName();
